requirements.txt
pytest==7.4.3
requests==2.31.0
allure-pytest==2.13.2
pytest-xdist==3.5.0
pytest-rerunfailures==13.0
PyYAML==6.0.1
jsonschema==4.20.0
pymysql==1.1.0
redis==5.0.1
Faker==22.0.0
cryptography==41.0.7
loguru==0.7.2
pydantic==2.5.3
python-dotenv==1.0.0
pytest-html==4.1.1
pytest-timeout==2.2.0
pytest-assume==2.4.3
